altera diretorio do arquivo streamer.py
